+++ Assets/VRM10/Editor/Components/Expression/ExpressionSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1d6be45aa67f5445884833ae12546bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
