+++ Assets/VRM10/Editor/Components/Expression/VRM10ControllerEditorExpression.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3387f1e8cba522b44a78609529674819
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
