+++ Assets/VRM10/Editor/Components/PropGui.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a55f4351fdc33f4fadbaa02abc5ffdb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
