+++ Assets/VRM10/Editor/Components/VRM10ControllerEditor.cs
@@ -1,61 +1,59 @@
-﻿using System.Collections.Generic;
-using UnityEditor;
+﻿using UnityEditor;
 using UnityEngine;
-using System.Linq;
 using System;
 
 namespace UniVRM10
 {
+    /// <summary>
+    /// VRM10Controller CustomEditor
+    /// 
+    /// * Controller
+    /// * Meta
+    /// * Expression
+    /// * LookAt
+    /// * ForstPerson
+    /// 
+    /// </summary>
     [CustomEditor(typeof(VRM10Controller))]
     public class VRM10ControllerEditor : Editor
     {
         VRM10Controller m_target;
-        SkinnedMeshRenderer[] m_renderers;
-        Dictionary<ExpressionKey, float> m_expressionKeyWeights = new Dictionary<ExpressionKey, float>();
 
-        public class ExpressionSlider
+        enum Tabs
         {
-            Dictionary<ExpressionKey, float> m_expressionKeys;
-            ExpressionKey m_key;
-
-            public ExpressionSlider(Dictionary<ExpressionKey, float> expressionKeys, ExpressionKey key)
-            {
-                m_expressionKeys = expressionKeys;
-                m_key = key;
-            }
-
-            public KeyValuePair<ExpressionKey, float> Slider()
-            {
-                var oldValue = m_expressionKeys[m_key];
-                var enable = GUI.enabled;
-                GUI.enabled = Application.isPlaying;
-                var newValue = EditorGUILayout.Slider(m_key.ToString(), oldValue, 0, 1.0f);
-                GUI.enabled = enable;
-                return new KeyValuePair<ExpressionKey, float>(m_key, newValue);
-            }
+            Controller,
+            Meta,
+            Expression,
+            LookAt,
+            FirstPerson,
         }
-        List<ExpressionSlider> m_sliders;
+        Tabs _tab = Tabs.Meta;
+
+        // for ScriptedObject
+        Editor m_metaEditor;
+
+        // expression
+        VRM10ControllerEditorExpression m_expression;
+
+        // for SerializedProperty
+        PropGui m_controller;
+        PropGui m_meta;
+        PropGui m_lookAt;
+        PropGui m_firstPerson;
+        PropGui m_asset;
 
         void OnEnable()
         {
             m_target = (VRM10Controller)target;
-
-            m_expressionKeyWeights = m_target.Expression.Clips.ToDictionary(x => ExpressionKey.CreateFromClip(x), x => 0.0f);
-            m_sliders = m_target.Expression.Clips
-                .Where(x => x != null)
-                .Select(x => new ExpressionSlider(m_expressionKeyWeights, ExpressionKey.CreateFromClip(x)))
-                .ToList()
-                ;
-
+            m_expression = new VRM10ControllerEditorExpression(m_target);
             if (m_target?.Meta.Meta != null)
             {
                 m_metaEditor = Editor.CreateEditor(m_target.Meta.Meta);
             }
-            m_controller = PropGui.FromObject(serializedObject, nameof(m_target.Controller));
-            m_meta = PropGui.FromObject(serializedObject, nameof(m_target.Meta));
-            m_expression = PropGui.FromObject(serializedObject, nameof(m_target.Expression));
-            m_lookAt = PropGui.FromObject(serializedObject, nameof(m_target.LookAt));
-            m_firstPerson = PropGui.FromObject(serializedObject, nameof(m_target.FirstPerson));
+            m_controller = new PropGui(serializedObject.FindProperty(nameof(m_target.Controller)));
+            m_meta = new PropGui(serializedObject.FindProperty(nameof(m_target.Meta)));
+            m_lookAt = new PropGui(serializedObject.FindProperty(nameof(m_target.LookAt)));
+            m_firstPerson = new PropGui(serializedObject.FindProperty(nameof(m_target.FirstPerson)));
         }
 
         void OnDisable()
@@ -69,67 +67,12 @@ namespace UniVRM10
             Tools.hidden = false;
         }
 
-        enum Tabs
-        {
-            Controller,
-            Meta,
-            Expression,
-            LookAt,
-            FirstPerson,
-        }
-        Tabs _tab = Tabs.Meta;
-
-        Editor m_metaEditor;
-
-        class PropGui
-        {
-            SerializedProperty m_prop;
-
-            PropGui(SerializedProperty property)
-            {
-                m_prop = property;
-            }
-
-            public static PropGui FromObject(SerializedObject serializedObject, string name)
-            {
-                var prop = serializedObject.FindProperty(name);
-                return new PropGui(prop);
-            }
-
-            public void RecursiveProperty()
-            {
-                var depth = m_prop.depth;
-                var iterator = m_prop.Copy();
-                for (var enterChildren = true; iterator.NextVisible(enterChildren); enterChildren = false)
-                {
-                    if (iterator.depth < depth)
-                        return;
-
-                    depth = iterator.depth;
-
-                    using (new EditorGUI.DisabledScope("m_Script" == iterator.propertyPath))
-                        EditorGUILayout.PropertyField(iterator, true);
-                }
-            }
-        }
-
-        PropGui m_controller;
-        PropGui m_meta;
-        PropGui m_expression;
-        PropGui m_lookAt;
-        PropGui m_firstPerson;
-        PropGui m_asset;
-
         public override void OnInspectorGUI()
         {
+            using (new EditorGUI.DisabledGroupScope(false))
             {
-                var backup = GUI.enabled;
-                GUI.enabled = true;
-
                 _tab = (Tabs)EditorGUILayout.EnumPopup("Select GUI", _tab);
                 EditorGUILayout.Separator();
-
-                GUI.enabled = backup;
             }
 
             serializedObject.Update();
@@ -150,8 +93,7 @@ namespace UniVRM10
                     break;
 
                 case Tabs.Expression:
-                    // m_expression.RecursiveProperty();
-                    ExpressionGUI();
+                    m_expression.OnGUI();
                     break;
 
                 case Tabs.LookAt:
@@ -164,39 +106,6 @@ namespace UniVRM10
             }
 
             serializedObject.ApplyModifiedProperties();
-        }
-
-        void ExpressionGUI()
-        {
-            EditorGUILayout.Space();
-
-            if (!Application.isPlaying)
-            {
-                EditorGUILayout.HelpBox("Enable when playing", MessageType.Info);
-            }
-
-            if (m_sliders != null)
-            {
-                EditorGUILayout.Space();
-                EditorGUILayout.LabelField("Expression Weights", EditorStyles.boldLabel);
-
-                var sliders = m_sliders.Select(x => x.Slider());
-                foreach (var slider in sliders)
-                {
-                    m_expressionKeyWeights[slider.Key] = slider.Value;
-                }
-                m_target.Expression.SetWeights(m_expressionKeyWeights);
-            }
-
-            EditorGUILayout.Space();
-            EditorGUILayout.LabelField("Override rates", EditorStyles.boldLabel);
-            EditorGUI.BeginDisabledGroup(true);
-            {
-                EditorGUILayout.Slider("Blink override rate", m_target.Expression.BlinkOverrideRate, 0f, 1f);
-                EditorGUILayout.Slider("LookAt override rate", m_target.Expression.LookAtOverrideRate, 0f, 1f);
-                EditorGUILayout.Slider("Mouth override rate", m_target.Expression.MouthOverrideRate, 0f, 1f);
-            }
-            EditorGUI.EndDisabledGroup();
         }
 
         enum VRMSceneUI
